git-gui: Change to display the combined diff in the case of conflicts.
The rationale for this change is that the previous behaviour did not allow
the user to make an informed decision about the likely consequences
of "Use Local Version", "Stage to commit" or "Use Remote Version" for conflicted files. The
reason for this is that in the conflicted case, successfully staged
remote hunks are effectively invisible to user (via the git-gui interface)
and hence cannot inform the user's decision making process.

For example, previously use of "Use Local Version" would silently discard
from the index, and the working tree any successfully merged remote hunks.
Since these hunks had never been displayed to the user, this
loss would be unnoticed and unexpected.

In the case of "Stage to commit", the successfully merged remote
hunks would be preserved in the index and the working tree but the fact
that there are successfully merged remote hunks would not be not
visible until after the "Stage to commit" action has been taken.
If the user did not check diff _after_ taking the action, the user
may have unwittingly commit changes from a remote hunk
that she was not aware of.

Similar coniderations also imply that the probable consequences
of "Use Remote Version" cannot be properly evaluated since the changes
in the local branch that would be undone by accepting the remote
branch's version of the file would are not visible to the user.

With this change, in the case of conflicted paths only, the git-gui diff window
displays the output of git diff -c. This output allows the user to properly
evaluate the consequences of the possible resolution actions.

Previous versions of this patch uses "diff HEAD" instead of "diff".
This version uses "diff -c" at the suggestion of Johannes Sixt.

Signed-off-by: Jon Seymour <jon.seymour@gmail.com>